Uncompiled Apps
Previously detection would fail if the user tried to push an application from source, as
no spring-core jar would be present in the source. This commit moves the search for the
jar from detect (which now only looks for a java app) to build, where the compiled app
is present.

Signed-off-by: Paul Harris <pharris@pivotal.io>